run_experiment1-kg_type.sh KGS ig, uig and euig; Added overall_comp_cost.

+++ util/clean_kge-k-means.sh
@@ -2,6 +2,13 @@
 
 #[kge-k-means_data]
 clean_kge-k-means() {
-    yes | rm ~/git/kg-summ-rec/docker/kge-k-means_data/application.log
-    yes | rm -r ~/git/kg-summ-rec/docker/kge-k-means_data/temp/*.*
+    if [ -f "$HOME/git/kg-summ-rec/docker/kge-k-means_data/application.log" ]
+    then 
+        yes | rm ~/git/kg-summ-rec/docker/kge-k-means_data/application.log
+    fi
+
+    if ls "$HOME/git/kg-summ-rec/docker/kge-k-means_data/temp/*.*" 1> /dev/null 2>&1
+    then
+        yes | rm -r ~/git/kg-summ-rec/docker/kge-k-means_data/temp/*.*
+    fi
 }
